pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = [
    {name = "Granth Agarwal",email = "heygranth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "blinker (==1.9.0)",
    "click (==8.1.8)",
    "flask (==3.1.0)",
    "greenlet (==3.1.1)",
    "hashids (==1.3.1)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.5)",
    "markupsafe (==3.0.2)",
    "python-dotenv (==1.0.1)",
    "qrcode (==8.0)",
    "sqlalchemy (==2.0.36)",
    "typing-extensions (==4.12.2)",
    "werkzeug (==3.1.3)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.1)",
    "certifi (==2024.12.14)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.1)",
    "cleo (==2.1.0)",
    "crashtest (==0.4.1)",
    "cryptography (==44.0.0)",
    "distlib (==0.3.9)",
    "dulwich (==0.22.7)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.16.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jeepney (==0.8.0)",
    "keyring (==25.6.0)",
    "more-itertools (==10.5.0)",
    "msgpack (==1.1.0)",
    "packaging (==24.2)",
    "pkginfo (==1.12.0)",
    "platformdirs (==4.3.6)",
    "poetry (==2.0.0)",
    "poetry-core (==2.0.0)",
    "pycparser (==2.22)",
    "pyproject-hooks (==1.2.0)",
    "rapidfuzz (==3.11.0)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "secretstorage (==3.3.3)",
    "setuptools (==75.7.0)",
    "shellingham (==1.5.4)",
    "tomlkit (==0.13.2)",
    "trove-classifiers (==2024.10.21.16)",
    "urllib3 (==2.3.0)",
    "virtualenv (==20.28.1)",
    "autopep8 (==2.3.1)",
    "docutils (==0.21.2)",
    "pycodestyle (==2.12.1)",
    "tomli-w (==1.1.0)",
    "yapf (==0.43.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
